test: stub core in cli tests
Our CLI tests test core functionality, most of which is duplicated
by our interface tests, consequently our tests take unnecessarily
long to run as the same code paths are being exercised by multiple
tests during a test run.

This PR refactors the tests to use sinon to stub core methods and
refocusses them on making sure we pass the correct arguments to
our core APIs.

The exceptions to this are where CLI tests run code that is not
part of the core interface API, such as `init` and starting/stopping
daemons.